+++ sopa/io/explorer/shapes.py
@@ -45,7 +45,11 @@ def pad_polygon(
 
 
 def write_polygons(
-    path: Path, polygons: Iterable[Polygon], max_vertices: int, is_dir: bool = True
+    path: Path,
+    polygons: Iterable[Polygon],
+    max_vertices: int,
+    is_dir: bool = True,
+    pixelsize: float = 0.2125,
 ) -> None:
     """Write a `cells.zarr.zip` file containing the cell polygonal boundaries
 
@@ -54,12 +58,13 @@ def write_polygons(
         polygons: A list of `shapely` polygons to be written
         max_vertices: The number of vertices per polygon (they will be transformed to have the right number of vertices)
         is_dir: If `False`, then `path` is a path to a single file, not to the Xenium Explorer directory.
+        pixelsize: Number of microns in a pixel. Invalid value can lead to inconsistent scales in the Explorer.
     """
     path = explorer_file_path(path, FileNames.SHAPES, is_dir)
 
     log.info(f"Writing {len(polygons)} cell polygons")
     coordinates = np.stack([pad_polygon(p, max_vertices) for p in polygons])
-    coordinates /= ExplorerConstants.MICRONS_TO_PIXELS
+    coordinates *= pixelsize
 
     num_cells = len(coordinates)
     cells_fourth = ceil(num_cells / 4)